Django auth final

+++ content/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
+from django.contrib.auth.decorators import login_required
 
 from .models import Content
-from user.models import User
 
 def index(request):
     all_posts = Content.objects.all()
@@ -9,29 +9,29 @@ def index(request):
         'posts': all_posts
     })
 
+@login_required
 def new_page(request):
-    users = User.objects.all()
-    return render(request, "content/new.html", context={
-        "users": users
-    })
+    return render(request, "content/new.html")
 
+@login_required
 def edit_page(request, post_id):
     post = Content.objects.get(id=post_id)
     return render(request, "content/edit.html", context={
         "post": post
     })
 
+@login_required
 def create(request):
     new_post = Content(
-        **request.POST
-        # title = request.POST['title'],
-        # text = request.POST['text'],
-        # user_id = request.POST['user_id'],
+        title = request.POST['title'],
+        text = request.POST['text'],
+        user_id = request.user.id,
     )
     new_post.save()
 
     return redirect("/")
 
+@login_required
 def update(request, post_id):
     post = Content.objects.get(id=post_id)
     post.title = request.POST['title']
@@ -40,8 +40,15 @@ def update(request, post_id):
 
     return redirect("/")
 
+@login_required
 def delete(request, post_id):
-    post = Content.objects.get(id=post_id)
-    post.delete()
+    current_user = request.user
+    current_user.load_posts()
+
+    user_post_ids = [p.id for p in current_user.post_list]
+
+    if int(post_id) in user_post_ids:
+        post = Content.objects.get(id=post_id)
+        post.delete()
 
     return redirect("/")
